requirements.txt
streamlit
opencv-python
Pillow
numpy
inference-sdk
pafy
python-dotenv 